patch: remove unnecessary confbridge patches
For oasis release, during any-call-to-any-as call distribution method,
conferences required some devicestates to detect the actual application
server that had the conference started. That doesn't happen in artemis
release, so this patches are no longer required.

Additionally, when a user hanguped its channel while the conference
is requesting the pin, CBAnn channels were not properly destroyed. This
problem was introduced with these patches, so it should no longer appear.